backend/requirements.txt
# Phase 0-1.5: Consolidated Dependencies for Local AI Agent Platform Backend

# Core Backend Dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.1
alembic==1.13.1

# AI & ML Dependencies
langchain==0.2.1
langgraph==0.0.60
# smol-agents is not on PyPI, will be treated as a local module or git dependency.
pyautogen==0.2.20
sentence-transformers==2.7.0
transformers==4.41.1
tokenizers==0.19.1
huggingface-hub==0.23.0
safetensors==0.4.3
openai==1.25.2
google-genai==1.0.0
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
jsonschema==4.22.0
python-docx==1.1.2
openpyxl==3.1.2
duckduckgo-search==5.3.1b1
pyperclip==1.8.2
Pillow==10.3.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
msal==1.28.0
aiohttp==3.9.5
faiss-cpu==1.8.0
ollama==0.2.0

# Automation, Monitoring & Settings Dependencies
apscheduler==3.10.4
psutil==5.9.8
gputil==1.4.0
watchdog==4.0.0
croniter==2.0.4
plotly==5.22.0
cryptography==42.0.7
keyring==25.2.1

# General Utilities
python-multipart==0.0.9
aiofiles==23.2.1
PyYAML==6.0.1
typing-extensions==4.11.0
matplotlib==3.8.4
lodash-py==0.2.1

# Development & Testing Dependencies
pytest==8.2.1
pytest-asyncio==0.23.6
black==24.4.2
ruff==0.4.4
mypy==1.10.0

# Windows-Specific Dependencies (some may be covered above)
pywin32==306

# MCP (Model Context Protocol) Dependencies
mcp==0.1.0
